Debugging simplification backend
