+++ src/CustomSymbolicUtilsSimplification.jl
@@ -10,72 +10,8 @@ catch
 end
 
 @from "Core.jl" import Options
-@from "InterfaceSymbolicUtils.jl" import SYMBOLIC_UTILS_TYPES
+@from "InterfaceSymbolicUtils.jl" import SYMBOLIC_UTILS_TYPES, multiply_powers
 @from "Utils.jl" import isgood, @return_on_false
-
-function multiply_powers(eqn::T)::Tuple{SYMBOLIC_UTILS_TYPES,Bool} where {T<:Union{<:Number,SymbolicUtils.Sym{<:Number}}}
-	return eqn, true
-end
-
-function multiply_powers(eqn::T, op::F)::Tuple{SYMBOLIC_UTILS_TYPES,Bool} where {F,T<:Union{SymbolicUtils.Term{<:Number},SymbolicUtils.Symbolic{<:Number}}}
-	args = SymbolicUtils.arguments(eqn)
-	nargs = length(args)
-	if nargs == 1
-        l, complete = multiply_powers(args[1])
-        @return_on_false complete eqn
-        @return_on_false isgood(l) eqn
-		return op(l), true
-	elseif op == ^
-		l, complete = multiply_powers(args[1])
-        @return_on_false complete eqn
-        @return_on_false isgood(l) eqn
-		n = args[2]
-        if typeof(n) <: Int
-            if n == 1
-                return l, true
-            elseif n == -1
-                return 1.0 / l, true
-            elseif n > 1
-                return reduce(*, [l for i=1:n]), true
-            elseif n < -1
-                return reduce(/, vcat([1], [l for i=1:abs(n)])), true
-            else
-                return 1.0, true
-            end
-        else
-            r, complete2 = multiply_powers(args[2])
-            @return_on_false complete2 eqn
-            return l ^ r, true
-        end
-	elseif nargs == 2
-        l, complete = multiply_powers(args[1])
-        @return_on_false complete eqn
-        @return_on_false isgood(l) eqn
-        r, complete2 = multiply_powers(args[2])
-        @return_on_false complete2 eqn
-        @return_on_false isgood(r) eqn
-		return op(l, r), true
-	else
-		# return mapreduce(multiply_powers, op, args)
-        # ## reduce(op, map(multiply_powers, args))
-        out = map(multiply_powers, args) #vector of tuples
-        for i=1:size(out, 1)
-            @return_on_false out[i][2] eqn
-            @return_on_false isgood(out[i][1]) eqn
-        end
-        cumulator = out[1][1]
-        for i=2:size(out, 1)
-            cumulator = op(cumulator, out[i][1])
-            @return_on_false isgood(cumulator) eqn
-        end
-        return cumulator, true
-	end
-end
-
-function multiply_powers(eqn::T)::Tuple{SYMBOLIC_UTILS_TYPES,Bool} where {T<:Union{SymbolicUtils.Term{<:Number},SymbolicUtils.Symbolic{<:Number}}}
-	op = SymbolicUtils.operation(eqn)
-	return multiply_powers(eqn, op)
-end
 
 # Operators required for each rule:
 function get_simplifier(binops::A, unaops::B) where {A,B}
@@ -161,12 +97,13 @@ function get_simplifier(binops::A, unaops::B) where {A,B}
                         ; kw...)
     end
     # reduce overhead of simplify by defining these as constant
-    serial_simplifier = If(istree, Fixpoint(default_simplifier()))
+    # serial_simplifier = If(istree, Fixpoint(default_simplifier()))
     serial_polynormal_simplifier = If(istree,
                                       Fixpoint(Chain((expand,
                                                       Fixpoint(default_simplifier())))))
     return serial_polynormal_simplifier
 end
+
 
 function custom_simplify(init_eqn::T, options::Options)::Tuple{SYMBOLIC_UTILS_TYPES, Bool} where {T<:SYMBOLIC_UTILS_TYPES}
     if !istree(init_eqn) #simplifier will return nothing if not a tree.
