+++ src/InterfaceSymbolicUtils.jl
@@ -81,6 +81,12 @@ function Base.convert(::typeof(Node), x::SymbolicUtils.Symbolic, options::Option
         return Node(String(x.name), varMap)
     end
 
+    # First, we remove integer powers:
+    y, good_return = multiply_powers(x)
+    if good_return
+        x = y
+    end
+
     op = convert_to_function(SymbolicUtils.operation(x))
     args = SymbolicUtils.arguments(x)
 
@@ -132,4 +138,68 @@ function symbolic_to_node(eqn::T, options::Options;
                           varMap::Union{Array{String, 1}, Nothing}=nothing)::Node where {T<:SymbolicUtils.Symbolic}
 
     convert(Node, eqn, options; varMap=varMap)
+end
+
+function multiply_powers(eqn::T)::Tuple{SYMBOLIC_UTILS_TYPES,Bool} where {T<:Union{<:Number,SymbolicUtils.Sym{<:Number}}}
+	return eqn, true
+end
+
+function multiply_powers(eqn::T, op::F)::Tuple{SYMBOLIC_UTILS_TYPES,Bool} where {F,T<:Union{SymbolicUtils.Term{<:Number},SymbolicUtils.Symbolic{<:Number}}}
+	args = SymbolicUtils.arguments(eqn)
+	nargs = length(args)
+	if nargs == 1
+        l, complete = multiply_powers(args[1])
+        @return_on_false complete eqn
+        @return_on_false isgood(l) eqn
+		return op(l), true
+	elseif op == ^
+		l, complete = multiply_powers(args[1])
+        @return_on_false complete eqn
+        @return_on_false isgood(l) eqn
+		n = args[2]
+        if typeof(n) <: Int
+            if n == 1
+                return l, true
+            elseif n == -1
+                return 1.0 / l, true
+            elseif n > 1
+                return reduce(*, [l for i=1:n]), true
+            elseif n < -1
+                return reduce(/, vcat([1], [l for i=1:abs(n)])), true
+            else
+                return 1.0, true
+            end
+        else
+            r, complete2 = multiply_powers(args[2])
+            @return_on_false complete2 eqn
+            return l ^ r, true
+        end
+	elseif nargs == 2
+        l, complete = multiply_powers(args[1])
+        @return_on_false complete eqn
+        @return_on_false isgood(l) eqn
+        r, complete2 = multiply_powers(args[2])
+        @return_on_false complete2 eqn
+        @return_on_false isgood(r) eqn
+		return op(l, r), true
+	else
+		# return mapreduce(multiply_powers, op, args)
+        # ## reduce(op, map(multiply_powers, args))
+        out = map(multiply_powers, args) #vector of tuples
+        for i=1:size(out, 1)
+            @return_on_false out[i][2] eqn
+            @return_on_false isgood(out[i][1]) eqn
+        end
+        cumulator = out[1][1]
+        for i=2:size(out, 1)
+            cumulator = op(cumulator, out[i][1])
+            @return_on_false isgood(cumulator) eqn
+        end
+        return cumulator, true
+	end
+end
+
+function multiply_powers(eqn::T)::Tuple{SYMBOLIC_UTILS_TYPES,Bool} where {T<:Union{SymbolicUtils.Term{<:Number},SymbolicUtils.Symbolic{<:Number}}}
+	op = SymbolicUtils.operation(eqn)
+	return multiply_powers(eqn, op)
 end
